requirements.txt
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain-text-splitters>=0.0.1
numpy>=1.24.3
loguru>=0.7.0
pillow>=10.0.0
pytesseract>=0.3.10
requests>=2.31.0
aiohttp>=3.8.5
asyncio>=3.4.3
tqdm>=4.66.1
pyairtable>=1.4.0 